fix(trusty-common): rank memory recall by similarity over importance (closes #633)
Before this fix, L1 (essential) drawers were scored by raw importance only
(up to 1.0). In `recall()` and `recall_deep()`, L1 results were inserted
first and then L2/L3 hits were appended — but there was no final re-ranking.
High-importance-but-irrelevant bulk-imported drawers (e.g. commit-log fixtures
at importance=1.0) dominated every cross-palace `memory_recall_all` result,
drowning out on-topic hits that scored 0.2–0.33 in the vector search.

Fix: after `retrieve_l2` / `retrieve_l3` runs, build a drawer_id→score map
from the vector-similarity results and call `rescore_l1_by_similarity`:
  - L1 drawers that appear in the HNSW results get their true cosine-
    similarity score, replacing the raw importance value.
  - L1 drawers NOT returned by the HNSW search are penalised by a
    `L1_NO_SIMILARITY_PENALTY = 0.15` coefficient so a max-importance
    off-topic drawer (0.15) is outranked by a mediocre-similarity L2 hit
    (e.g. 0.5 * 0.4 = 0.20). Importance remains a tiebreaker, not the
    primary key.
  - The merged list is then sorted by score descending before returning.

Also splits the large tests inline in retrieval.rs into a separate
`retrieval/tests.rs` file (retrieval.rs → retrieval/mod.rs + tests.rs)
to satisfy the line-cap ratchet after adding the new ordering tests;
updates `.line-cap-allowlist.tsv` accordingly.

No per-source anti-domination (diversity constraint) is implemented in this
PR — the similarity-first ranking already addresses the reported symptom, and
source-based diversity is left for a follow-up issue.
